% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_api_players.R
\name{extract_player_info}
\alias{extract_player_info}
\title{Extract player information into a structured tibble}
\usage{
extract_player_info(players, db_path = NULL)
}
\arguments{
\item{players}{Named list of player data as returned by load_all_sleeper_players()}

\item{db_path}{Character string specifying the path to save the SQLite database, NULL to skip saving}
}
\value{
A tibble containing extracted player information
}
\description{
Extracts player information from a nested list structure into a structured tibble
}
\examples{
\dontrun{
path_player_data <- "./data/raw/sleeper_players/sleeper_player_data_2025_MAY_14.json"
players <- load_all_sleeper_players(path_player_data)
path_player_data_parsed <- "./data/parsed/sleeper_all_players"
players_extracted <- extract_player_info(players, path_player_data_parsed)
}
}
