% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_api_drafts.R
\name{get_all_draft_picks}
\alias{get_all_draft_picks}
\title{Get all draft picks for a specific draft}
\usage{
get_all_draft_picks(draft_id)
}
\arguments{
\item{draft_id}{Character. The unique identifier for the draft}
}
\value{
A list containing all picks made in the specified draft
}
\description{
Retrieves all picks made in a specific draft
}
\examples{
\dontrun{
draft_picks <- get_all_draft_picks("123456789")
}
}
