% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_api_leagues.R
\name{get_league_rosters}
\alias{get_league_rosters}
\title{Get league rosters information}
\usage{
get_league_rosters(league_id)
}
\arguments{
\item{league_id}{Character string representing the league ID}
}
\value{
A tibble containing rosters information
}
\description{
Fetches information about all rosters in a specific league from the Sleeper API
}
\examples{
\dontrun{
rosters_data <- get_league_rosters("123456789")
}
}
