% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_api_leagues.R
\name{get_league_users}
\alias{get_league_users}
\title{Get league users information}
\usage{
get_league_users(league_id)
}
\arguments{
\item{league_id}{Character string representing the league ID}
}
\value{
A tibble containing users information
}
\description{
Fetches information about all users in a specific league from the Sleeper API
}
\examples{
\dontrun{
users_data <- get_league_users("123456789")
}
}
