% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_api_players.R
\name{get_sleeper_player_vars}
\alias{get_sleeper_player_vars}
\title{Get Sleeper player variable names}
\usage{
get_sleeper_player_vars()
}
\value{
A character vector of variable names
}
\description{
Returns a character vector of variable names available in the Sleeper player data
}
\examples{
vars <- get_sleeper_player_vars()
}
