% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_api_drafts.R
\name{get_specific_draft}
\alias{get_specific_draft}
\title{Get data for a specific draft}
\usage{
get_specific_draft(draft_id)
}
\arguments{
\item{draft_id}{Character. The unique identifier for the draft}
}
\value{
A list containing data for the specified draft
}
\description{
Retrieves metadata and information about a specific draft
}
\examples{
\dontrun{
draft_data <- get_specific_draft("123456789")
}
}
