% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_api_leagues.R
\name{get_specific_league}
\alias{get_specific_league}
\title{Get specific league information}
\usage{
get_specific_league(league_id)
}
\arguments{
\item{league_id}{Character string representing the league ID}
}
\value{
A tibble containing league information
}
\description{
Fetches information about a specific league from the Sleeper API
}
\examples{
\dontrun{
league_data <- get_specific_league("123456789")
}
}
