% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_api_users.R
\name{get_user}
\alias{get_user}
\title{Retrieve User Information from Sleeper API}
\usage{
get_user(username = NULL, user_id = NULL)
}
\arguments{
\item{username}{Character string. The Sleeper username to look up. Default is NULL.}

\item{user_id}{Character string. The Sleeper user ID to look up. Default is NULL.}
}
\value{
A list containing user information returned by the Sleeper API.
}
\description{
Fetches user data from the Sleeper API using either a username or user ID.
This function returns the full user object containing profile information.
}
\details{
Exactly one of username or user_id must be provided. The function will make
an HTTP GET request to the appropriate Sleeper API endpoint and return the
parsed JSON response.
}
\examples{
\dontrun{
# Get user by username
user_info <- get_user(username = "sleeper_username")

# Get user by user_id
user_info <- get_user(user_id = "12345678")
}

}
