% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_api_players.R
\name{load_all_sleeper_players}
\alias{load_all_sleeper_players}
\title{Load Sleeper NFL player data from JSON file}
\usage{
load_all_sleeper_players(file_path)
}
\arguments{
\item{file_path}{Character string specifying the full path to the JSON file}
}
\value{
A named list containing complete player data
}
\description{
Loads previously saved Sleeper NFL player data from a JSON file
}
\examples{
\dontrun{
path_player_data <- "./data/raw/sleeper_players/sleeper_player_data_2025_MAY_14.json"
players <- load_all_sleeper_players(path_player_data)
}
}
