% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_api_players.R
\name{load_or_fetch_sleeper_players}
\alias{load_or_fetch_sleeper_players}
\title{Load existing player data or fetch new data if not available}
\usage{
load_or_fetch_sleeper_players(save_path = "./data/raw/sleeper_players")
}
\arguments{
\item{save_path}{Character string specifying the directory path to save or look for player data}
}
\value{
A named list containing complete player data
}
\description{
Checks for existing player data for the current date and either loads it or fetches new data
}
\examples{
\dontrun{
path_player_data <- "./data/raw/sleeper_players"
players <- load_or_fetch_sleeper_players(path_player_data)
}
}
