% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logging.R
\name{log_debug}
\alias{log_debug}
\title{Log debug message}
\usage{
log_debug(message, .envir = parent.frame())
}
\arguments{
\item{message}{Character string with the log message (can use glue syntax)}

\item{.envir}{Environment for glue evaluation}
}
\description{
Log debug message
}
