% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logging.R
\name{log_error}
\alias{log_error}
\title{Log error message (will stop execution)}
\usage{
log_error(message, .envir = parent.frame())
}
\arguments{
\item{message}{Character string with the log message (can use glue syntax)}

\item{.envir}{Environment for glue evaluation}
}
\description{
Log error message (will stop execution)
}
