% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logging.R
\name{log_info}
\alias{log_info}
\title{Log info message}
\usage{
log_info(message, .envir = parent.frame())
}
\arguments{
\item{message}{Character string with the log message (can use glue syntax)}

\item{.envir}{Environment for glue evaluation}
}
\description{
Log info message
}
