% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logging.R
\name{set_log_colors}
\alias{set_log_colors}
\title{Enable or disable colored logging output}
\usage{
set_log_colors(use_colors = TRUE)
}
\arguments{
\item{use_colors}{Logical indicating whether to use colors}
}
\description{
Enable or disable colored logging output
}
