% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sleeper_api.R
\name{status_sleeper_api}
\alias{status_sleeper_api}
\title{Check status of Sleeper API response}
\usage{
status_sleeper_api(response)
}
\arguments{
\item{response}{httr response object}
}
\value{
NULL if successful, stops execution if error
}
\description{
Check status of Sleeper API response
}
\keyword{internal}
